tools/stress-test/requirements.txt
# DarkAges Stress Test Tools - Python Dependencies
# Python 3.8+ required

# Core dependencies (standard library only for basic operation)
# asyncio, socket, struct, json, time, random, dataclasses, typing, pathlib

# Required for integration testing (WP-6-5)
redis>=4.5.0           # Redis client for hot-state validation
cassandra-driver>=3.28.0  # ScyllaDB/Cassandra client for persistence testing

# Optional development dependencies:
pytest>=7.0.0          # For running unit tests
pytest-asyncio>=0.21.0 # Async test support
black>=22.0.0          # Code formatting
mypy>=0.950            # Type checking
flake8>=4.0.0          # Linting

# Optional advanced features:
flatbuffers>=2.0       # For full FlatBuffers serialization (instead of simplified binary)
numpy>=1.21.0          # For advanced statistical analysis
matplotlib>=3.5.0      # For plotting results
docker>=6.0.0          # For chaos testing with container control

# Network chaos testing (Linux only - tc is used via subprocess)
# iproute2 package required on host system
